Store waitForBalance in test utils
