+++ ui/core/src/test/utils/sleep.ts
@@ -1,0 +1,2 @@
+export const sleep = (ms: number) =>
+  new Promise((done) => setTimeout(done, ms));
